Yolov8/requirements.txt
streamlit
opencv-python
pillow
numpy
ultralytics
